requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.33.0
django-filter==1.0.4
django-webpack-loader==0.5.0
djangorestframework==3.6.4
idna==2.6
Markdown==2.6.9
mysqlclient==1.3.12
oauthlib==2.0.3
pkg-resources==0.0.0
pymongo==3.5.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
